Fix tree-shaking issues related to the `useSuspenseQuery` file
